requirement.txt
langchain
langchain-google-genai
python-dotenv
colorama
watchdog
plyer  # Optional for desktop notifications

# Dev only
twine
build
